Add AngularJS API docs for built-in models
Add apidocs/describe-builtin-models.js - a script that sets
up a loopback app with all built-in models attached and generates
the AngularJS services for this app. The generated ngdoc comments
are modified to make them compatible with strong-docs' parser.

Add a `prepublish` script to run the generator before publishing
to npmjs.org. Add `apidocs/.npmignore` to override `apidocs/.gitignore`
and include the generated file in the npm package.

Add docs.json configured to run the new script and use the generated
services file as a source for strong-docs.

When the generator script is run by apidocs.strongloop.com, the module
dependencies are not installed. The script detects this situation and
exits immediately, since the output file is already included in the
npm package.