Send auth header to all URLs on the same host
Modify the request interceptor to insert auth header on all requests
matching the same host as urlBase. The host includes the port,
if supplied.

If no host is supplied for urlBase (e.g. "/api") then location.host
is used for the comparison. All requests with no host automatically
pass and get the auth header inserted.

This allows other endpoints of the loopback app (not in urlBase)
to receive the AccessToken.